% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lineworkmaps-package.r
\docType{package}
\name{lineworkmaps}
\alias{lineworkmaps}
\alias{lineworkmaps-package}
\title{Compilation of \href{http://projectlinework.org/}{Project Linework} maps in
\code{SpatialPolygonsDataFrame} and \code{ggplot2} fortified objects ready for mapping}
\description{
The following lineworks were created, curated or edited by the following:

\itemize{
\item \emph{Elmer Casual} - Marty Elmer
\item \emph{Charmingly Inaccurate} - Daniel P Huffman
\item \emph{Geo Metro}- Daniel P Huffman
\item \emph{Moriarty Hand}- Dylan Moriarty
\item \emph{Times Approximate} - Daniel P Huffman
\item \emph{Twenty Seventy} - Marty Elmer
\item \emph{Wargames} - Daniel P Huffman
\item \emph{Weekend Update} - Jonah Adkins
}
}
\author{
Bob Rudis (@hrbrmstr)
}

