% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linework.R
\name{lineworks}
\alias{lineworks}
\title{Return vector of available linework shapes and optionally print verbose
description of shapfile data elements.}
\usage{
lineworks(verbose = FALSE)
}
\arguments{
\item{verbose}{return information about the structure of the additional
shapefile data? (Default: \code{FALSE} - just return names of
lineworks.)}
}
\value{
character vector
}
\description{
Return vector of available linework shapes and optionally print verbose
description of shapfile data elements.
}

