% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linework.R
\docType{data}
\name{linework}
\alias{linework}
\title{Project Linework spatial & fortified objects}
\format{A list with six named objects and 24 named nested objects}
\usage{
data(linework)
}
\description{
Project Linework spatial & fortified objects
}
\note{
Last updated 2015-07-20.
}
\keyword{datasets}

